requirements.txt
streamlit>=1.44
pandas>=2.2
numpy>=1.26
scipy>=1.12
scikit-learn>=1.5
plotly>=5.24
openpyxl>=3.1
umap-learn>=0.5
